backend/requirements.txt
fastapi
uvicorn[standard]
jinja2
livekit-api
pydantic-settings
python-multipart